+++ src/core/modules/market/DealFactory.sol
@@ -4,10 +4,12 @@ pragma solidity ^0.8.19;
 
 import "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
-import "src/utils/OwnableUpgradableDiamond.sol";
+
+import "src/core/modules/BaseModule.sol";
+
 import "src/deal/interfaces/IDeal.sol";
 import "src/deal/DealProxy.sol";
-import "src/core/modules/BaseModule.sol";
+
 import "./interfaces/IDealFactory.sol";
 
 /*