+++ test/utils/TestWithDeployment.sol
@@ -1,32 +1,29 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.19;
 
-import {Test, console2} from "forge-std/Test.sol";
-import "forge-std/console.sol";
+import {Test} from "forge-std/Test.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "@openzeppelin/contracts/proxy/ERC1967/ERC1967Proxy.sol";
-import "src/deal/interfaces/IDeal.sol";
-import "src/dev/TestERC20.sol";
-import "src/core/Core.sol";
-import "src/deal/Deal.sol";
-import "src/core/modules/market/Market.sol";
-import "src/core/modules/market/DealFactory.sol";
+
+import "src/core/interfaces/ICore.sol";
+import "src/core/interfaces/ICapacityConst.sol";
+import "src/core/modules/capacity/interfaces/ICapacity.sol";
 import "src/core/modules/market/interfaces/IMarket.sol";
 import "src/core/modules/market/interfaces/IDealFactory.sol";
-import "src/core/modules/capacity/Capacity.sol";
-import "src/core/modules/capacity/interfaces/ICapacity.sol";
+import "src/deal/interfaces/IDeal.sol";
 
-library DeployDealSystem {
+import {PRECISION} from "src/utils/Common.sol";
+
+contract TestWithDeployment is Test {
     // ------------------ Types ------------------
     struct Deployment {
         bool initialized;
         IERC20 tUSD;
         IDeal dealImpl;
-        Core core;
-        Market market;
-        Capacity capacity;
-        DealFactory dealFactory;
+        ICore core;
+        IMarket market;
+        ICapacity capacity;
+        IDealFactory dealFactory;
     }
 
     // ------------------ Constants ------------------
@@ -53,22 +50,23 @@ library DeployDealSystem {
     bytes32 public constant DEFAULT_DIFFICULTY_TARGET =
         0x00000000ffffffffffffffffffffffffffffffffffffffffffffffffffffffff;
     bytes32 public constant DEFAULT_INIT_GLOBAL_NONCE = keccak256("init_global_nonce");
+    address public DEFAULT_ORACLE = address(456);
 
     // ------------------ Variables ------------------
-    function deployDealSystem() internal returns (Deployment memory deployment) {
-        deployment.tUSD = IERC20(new TestERC20("Test USD", "tUSD", 8));
+    Deployment public deployment;
 
-        Deal dealImpl = new Deal();
-        Core coreImpl = new Core();
+    function _deploySystem() internal {
+        deployment.tUSD = IERC20(deployCode("out/TestERC20.sol/TestERC20.json", abi.encode("Test USD", "tUSD", 8)));
 
-        deployment.dealImpl = dealImpl;
+        IDeal dealImpl = IDeal(deployCode("out/Deal.sol/Deal.json"));
+        ICore coreImpl = ICore(deployCode("out/Core.sol/Core.json"));
 
-        deployment.core = Core(
+        ICore core = ICore(
             address(
                 new ERC1967Proxy(
                     address(coreImpl),
                     abi.encodeWithSelector(
-                        Core.initialize.selector,
+                        ICore.initialize.selector,
                         DEFAULT_EPOCH_DURATION,
                         DEFAULT_MIN_DEPOSITED_EPOCHS,
                         DEFAULT_MIN_REMATCHING_EPOCHS,
@@ -76,55 +74,63 @@ library DeployDealSystem {
                         DEFAULT_MAX_PROTOCOL_VERSION,
                         dealImpl,
                         DEFAULT_IS_WHITELIST_ENABLED,
-                        DEFAULT_FLT_PRICE,
-                        DEFAULT_USD_COLLATERAL_PER_UNIT,
-                        DEFAULT_USD_TARGET_REVENUE_PER_EPOCH,
-                        DEFAULT_MIN_DURATION,
-                        DEFAULT_MIN_REWARD_PER_EPOCH,
-                        DEFAULT_MAX_REWARD_PER_EPOCH,
-                        DEFAULT_VESTING_PERIOD_DURATION,
-                        DEFAULT_VESTING_PERIOD_COUNT,
-                        DEFAULT_SLASHING_RATE,
-                        DEFAULT_MIN_REQUIERD_PROOFS_PER_EPOCH,
-                        DEFAULT_MAX_PROOFS_PER_EPOCH,
-                        DEFAULT_WITHDRAW_EPOCHS_AFTER_FAILED,
-                        DEFAULT_MAX_FAILED_RATIO,
-                        DEFAULT_DIFFICULTY_TARGET,
-                        DEFAULT_INIT_REWARD_POOL,
-                        new RandomXProxy()
+                        ICapacityConst.CapacityConstInitArgs({
+                            fltPrice: DEFAULT_FLT_PRICE,
+                            usdCollateralPerUnit: DEFAULT_USD_COLLATERAL_PER_UNIT,
+                            usdTargetRevenuePerEpoch: DEFAULT_USD_TARGET_REVENUE_PER_EPOCH,
+                            minDuration: DEFAULT_MIN_DURATION,
+                            minRewardPerEpoch: DEFAULT_MIN_REWARD_PER_EPOCH,
+                            maxRewardPerEpoch: DEFAULT_MAX_REWARD_PER_EPOCH,
+                            vestingPeriodDuration: DEFAULT_VESTING_PERIOD_DURATION,
+                            vestingPeriodCount: DEFAULT_VESTING_PERIOD_COUNT,
+                            slashingRate: DEFAULT_SLASHING_RATE,
+                            minProofsPerEpoch: DEFAULT_MIN_REQUIERD_PROOFS_PER_EPOCH,
+                            maxProofsPerEpoch: DEFAULT_MAX_PROOFS_PER_EPOCH,
+                            withdrawEpochsAfterFailed: DEFAULT_WITHDRAW_EPOCHS_AFTER_FAILED,
+                            maxFailedRatio: DEFAULT_MAX_FAILED_RATIO,
+                            difficulty: DEFAULT_DIFFICULTY_TARGET,
+                            initRewardPool: DEFAULT_INIT_REWARD_POOL,
+                            randomXProxy: deployCode("out/RandomXProxy.sol/RandomXProxy.json"),
+                            oracle: DEFAULT_ORACLE
+                        })
                     )
                 )
             )
         );
 
-        deployment.market = Market(
+        deployment.core = core;
+        deployment.dealImpl = dealImpl;
+
+        deployment.market = IMarket(
             address(
                 new ERC1967Proxy(
-                    address(new Market(deployment.core)), abi.encodeWithSelector(Market.initialize.selector, dealImpl)
+                    deployCode("out/Market.sol/Market.json", abi.encode(address(core))),
+                    abi.encodeWithSelector(IMarket.initialize.selector, dealImpl)
                 )
             )
         );
 
-        deployment.capacity = Capacity(
+        deployment.capacity = ICapacity(
             payable(
                 address(
                     new ERC1967Proxy(
-                        address(new Capacity(deployment.core)),
-                        abi.encodeWithSelector(Capacity.initialize.selector, DEFAULT_INIT_GLOBAL_NONCE)
+                        deployCode("out/Capacity.sol/Capacity.json", abi.encode(address(core))),
+                        abi.encodeWithSelector(ICapacity.initialize.selector, DEFAULT_INIT_GLOBAL_NONCE)
                     )
                 )
             )
         );
 
-        deployment.dealFactory = DealFactory(
+        deployment.dealFactory = IDealFactory(
             address(
                 new ERC1967Proxy(
-                    address(new DealFactory(deployment.core)), abi.encodeWithSelector(DealFactory.initialize.selector)
+                    deployCode("out/DealFactory.sol/DealFactory.json", abi.encode(address(core))),
+                    abi.encodeWithSelector(IDealFactory.initialize.selector)
                 )
             )
         );
 
-        deployment.core.initializeModules(deployment.capacity, deployment.market, deployment.dealFactory);
+        core.initializeModules(deployment.capacity, deployment.market, deployment.dealFactory);
 
         deployment.initialized = true;
     }