+++ ts-client/src/dealExplorerClient/indexerClient/queries/providers-query.generated.ts
@@ -15,7 +15,7 @@ export type ProvidersQueryQueryVariables = Types.Exact<{
 }>;
 
 
-export type ProvidersQueryQuery = { __typename?: 'Query', providers: Array<{ __typename?: 'Provider', id: string, name: string, createdAt: any, computeUnitsAvailable: number, computeUnitsTotal: number, approved: boolean, offers?: Array<{ __typename?: 'Offer', id: string, createdAt: any, pricePerEpoch: any, computeUnitsTotal?: number | null, computeUnitsAvailable?: number | null, paymentToken: { __typename?: 'Token', id: string, symbol: string, decimals: number }, effectors?: Array<{ __typename?: 'OfferToEffector', effector: { __typename?: 'Effector', id: string, description: string } }> | null, provider: { __typename?: 'Provider', id: string }, peers?: Array<{ __typename?: 'Peer', id: string }> | null }> | null }>, graphNetworks: Array<{ __typename?: 'GraphNetwork', providersTotal: any }> };
+export type ProvidersQueryQuery = { __typename?: 'Query', providers: Array<{ __typename?: 'Provider', id: string, name: string, createdAt: any, computeUnitsAvailable: number, computeUnitsTotal: number, approved: boolean, offers?: Array<{ __typename?: 'Offer', id: string, createdAt: any, pricePerEpoch: any, computeUnitsTotal?: number | null, computeUnitsAvailable?: number | null, paymentToken: { __typename?: 'Token', id: string, symbol: string, decimals: number }, effectors?: Array<{ __typename?: 'OfferToEffector', effector: { __typename?: 'Effector', id: string, description: string } }> | null, provider: { __typename?: 'Provider', id: string }, peers?: Array<{ __typename?: 'Peer', id: string }> | null }> | null }>, graphNetworks: Array<{ __typename?: 'GraphNetwork', providersRegisteredTotal: any }> };
 
 export type ProviderOfProvidersQueryFragment = { __typename?: 'Provider', id: string, name: string, createdAt: any, computeUnitsAvailable: number, computeUnitsTotal: number, approved: boolean, offers?: Array<{ __typename?: 'Offer', id: string, createdAt: any, pricePerEpoch: any, computeUnitsTotal?: number | null, computeUnitsAvailable?: number | null, paymentToken: { __typename?: 'Token', id: string, symbol: string, decimals: number }, effectors?: Array<{ __typename?: 'OfferToEffector', effector: { __typename?: 'Effector', id: string, description: string } }> | null, provider: { __typename?: 'Provider', id: string }, peers?: Array<{ __typename?: 'Peer', id: string }> | null }> | null };
 
@@ -68,7 +68,7 @@ export const ProvidersBasicOfferFragmentDoc = gql`
   provider {
     id
   }
-  peers {
+  peers(where: {deleted: false}) {
     id
   }
 }
@@ -94,7 +94,7 @@ export const ProvidersQueryDocument = gql`
     ...ProviderOfProvidersQuery
   }
   graphNetworks(first: 1) {
-    providersTotal
+    providersRegisteredTotal
   }
 }
     ${ProviderOfProvidersQueryFragmentDoc}`;